==> notebooks/00_setup_environment.ipynb <==
# 00_setup_environment.ipynb

# Install dependencies from requirements.txt

!pip install -r ../requirements.txt 

# Verify Imports
import sys
import numpy as np
import pandas as pd
import tqdm
import streamlit as st
import gdown

print("Python:", sys.version)
print("NumPy:", np.__version__)
print("Pandas:", pd.__version__)
print("tqdm:", tqdm.__version__)
print("Streamlit:", st.__version__)
print("gdown:", gdown.__version__)
